requirements.txt
# Quantum Network Simulation
qunetsim

# Post-Quantum Cryptography
liboqs-python

# Symmetric Encryption
pycryptodome

# Additional utilities
numpy
matplotlib
pytest
